Implement the method to check if the logged user is an admin
