+++ src/app/app.module.ts
@@ -41,7 +41,8 @@ import { DatastoreService } from 'app/services/datastore.service';
 import { CountriesService } from 'app/services/countries.service';
 import { RegisterComponent } from 'app/pages/register/register.component';
 import { ObservationListComponent } from 'app/pages/observations/observation-list.component';
-import { AuthService, TokenService } from 'app/services/auth.service';
+import { AuthService } from 'app/services/auth.service';
+import { TokenService } from 'app/services/token.service';
 import { LoginComponent } from 'app/pages/login/login.component';
 import { AppComponent } from 'app/app.component';
 import { AppRoutingModule } from 'app/app-routing.module';
